fixed deserialization bug

+++ Lemon/src/serialization/basic_types_serializer.cpp
@@ -99,7 +99,7 @@ void deserialize(u64& value, const Value& iter)
 }
 void deserialize(f32& value, const Value& iter)
 {
-    value = iter.GetFloat();
+    value = iter.GetDouble();
 }
 void deserialize(f64& value, const Value& iter)
 {
@@ -112,17 +112,17 @@ void deserialize(string& value, const Value& iter)
 void deserialize(vec2& value, const Value& iter)
 {
     auto a = iter.GetArray();
-    value  = { a[0].GetFloat(), a[1].GetFloat() };
+    value  = { a[0].GetDouble(), a[1].GetDouble() };
 }
 void deserialize(vec3& value, const Value& iter)
 {
     auto a = iter.GetArray();
-    value  = { a[0].GetFloat(), a[1].GetFloat(), a[2].GetFloat() };
+    value  = { a[0].GetDouble(), a[1].GetDouble(), a[2].GetDouble() };
 }
 void deserialize(vec4& value, const Value& iter)
 {
     auto a = iter.GetArray();
-    value  = { a[0].GetFloat(), a[1].GetFloat(), a[2].GetFloat(), a[3].GetFloat() };
+    value  = { a[0].GetDouble(), a[1].GetDouble(), a[2].GetDouble(), a[3].GetDouble() };
 }
 void deserialize(mat4& value, const Value& iter)
 {
@@ -133,10 +133,10 @@ void deserialize(mat4& value, const Value& iter)
     auto e = a[3].GetArray();
 
     value = {
-        { b[0].GetFloat(), b[1].GetFloat(), b[2].GetFloat(), b[3].GetFloat() },
-        { c[0].GetFloat(), c[1].GetFloat(), c[2].GetFloat(), c[3].GetFloat() },
-        { d[0].GetFloat(), d[1].GetFloat(), d[2].GetFloat(), d[3].GetFloat() },
-        { e[0].GetFloat(), e[1].GetFloat(), e[2].GetFloat(), e[3].GetFloat() }
+        { b[0].GetDouble(), b[1].GetDouble(), b[2].GetDouble(), b[3].GetDouble() },
+        { c[0].GetDouble(), c[1].GetDouble(), c[2].GetDouble(), c[3].GetDouble() },
+        { d[0].GetDouble(), d[1].GetDouble(), d[2].GetDouble(), d[3].GetDouble() },
+        { e[0].GetDouble(), e[1].GetDouble(), e[2].GetDouble(), e[3].GetDouble() }
     };
 }
 void deserialize(string_id& value, const Value& iter)
